Reorganize python code
