+++ python/tests/test_backbone_inmemory.py
@@ -2,13 +2,13 @@ from unittest.mock import Mock
 import pytest
 import stix2
 
-import vast.backbones
+import vast.fabric.backbones
 
 
 @pytest.mark.asyncio
 async def test_backbone_pub_sub():
     callback = Mock()
-    backbone = vast.backbones.InMemory()
+    backbone = vast.fabric.backbones.InMemory()
     message = stix2.Bundle()
     topic = "test"
     await backbone.subscribe(topic, callback)
