+++ python/tests/test_vast.py
@@ -1,23 +1,6 @@
-# TODO: we'll revamp this later.
-# import apps.test as test
-# import asyncio
-# import pyvast
-# from backbone import InMemoryBackbone
-# from fabric import Fabric
-#
-# import utils.asyncio
-# import utils.config
-# import utils.logging
-#
-# logger = utils.logger.create("test")
-#
-# def main():
-#    config = utils.config.parse()
-#    utils.logging.configure(config.logging)
-#    utils.asyncio.run_forever(start(config))
-#
-#
-#
-# if __name__ == "__main__":
-#    asyncio.run(main())
-#    # asyncio.run(test())
+from vast import VAST
+
+def test_vast():
+    v = VAST()
+    # TODO: make sure we have a vast binary nearby
+    assert(True)
