+++ python/vast/__init__.py
@@ -1,3 +1,5 @@
-from .backbone import Backbone
-from .fabric import Fabric
 from .vast import VAST
+import vast.fabric.apps
+import vast.fabric.backbones
+import vast.utils.config
+import vast.utils.logging
