+++ python/vast/vast/__init__.py
@@ -1,0 +1,1 @@
+from .vast import VAST
